+++ packages/twopoint5d/src/stage/IStage.ts
@@ -5,7 +5,10 @@ import type {ThreeRendererType} from '../display/types.js';
 export type RenderCmdFunc = (scene: Scene, camera: Camera, autoClear: boolean) => void;
 
 export interface IStage {
+  name: string;
+
   resize(width: number, height: number): void;
   renderFrame(renderer: ThreeRendererType, now: number, deltaTime: number, frameNo: number, renderCmd?: RenderCmdFunc): void;
+
   getRenderPass?(): RenderPass;
 }